Add GitHub Workflow for Docker build and push
Created a new GitHub workflow 'publish.yml' in the '.github/workflows' directory to automate the building and pushing of Docker images on pushes to the 'master' branch. This will expedite the process of publishing new changes to the Docker image by automating the process. The workflow runs on Ubuntu, checks out the repository, logs into the GitHub Container Registry (ghcr.io), extracts Docker metadata, then builds and pushes the Docker image.