Replace minio URL property with public image URL
The file 'ImageServiceImpl.java' was updated to replace the value of 'minioBucketPath' from spring.minio.url to PUBLIC_IMAGE_URL. It allows to reference resources from the publicly accessible URL. This change was done for the flexibility to use a different URL for public access to images. The accompanying 'application-docker.properties' file was added with a PUBLIC_IMAGE_URL variable to support the change.